#2422 Consistently allow transitions to tiny links from everywhere.
It does not matter whether we open tiny link from some centre / master, module menu or main menu. This transition should always perform opening of Entity Master for NEW instance exactly as for PERSISTED '/master/...' links.

Whether the user can move back to that tiny link is a separate matter (spoiler, they should not be able to do that).

 * Abstracted out the same logic for animated / non-animated transitions (+jsDocs).
 * Used both '/tiny/...' and '/master/...' links there for Entity Masters opening.
 * Taken care of reducing the chance of errors by using 'getCurrentlySelectedElement' lazy loading.
 * Added missing spaces between comment delimiters and comments.
 * Converted some docs to Markdown style, even though it is a JS file, for consistency with Java.
 * Added missing documentation to not easily comprehensible execution paths.