#2109 Rationalise a test assertion
Asserting that a specific number of tokens is retrieved will not help in detecting erronous situations.
For example, if the retrieval logic erronously skips some tokens.

Although, this assertion would be helpful if platform developers
consistently adjusted the expected number of tokens as part of the
change management process.
But, in absence of automated verification of changes this is not practical.